Added upload handler.
Test currently fails, because the handler is triggered too early. When the upload command has been initiated, the currentCommand is set to upload but at this point, the login command may not have returned. When it does, its response is sent to the upload handler. There needs to be a command queue.