Fixes #10. (Yet Another Attempt.)
When there was no connection at all, the connection error handler would bail out before the update method was called on DbcConnector and therefore no imageId was being computed and returned. Therefore the error handler at the very top got a nil imageId and could not reset the cell to timeout state.